Move drag selection to a helper class
